fixed mars

+++ mars/mars.py
@@ -12,17 +12,19 @@ mars = Blueprint('mars', __name__,
     template_folder='mars_templates', static_folder='mars_static')
 
 @mars.route('/mars')
-def home():
-    mars = collection.find_one()
-    if mars == None:
+def mars_home():
+    mars_data = collection.find_one()
+    if mars_data == None:
         return render_template('204.html')
-    return render_template('mars_index.html', mars=mars)
+    return render_template('mars_index.html', mars=mars_data)
 
 
 @mars.route('/mars/scrape')
 def call_scrape():
-    collection.drop()
-    collection.insert_one(scrape())
+    mars_data = scrape()
+    if mars_data:
+        collection.drop()
+        collection.insert_one(scrape())
     return redirect('/mars', code=302)
 
 if __name__ == '__main__':
