+++ mars/scrape/scrape_to_db.py
@@ -7,6 +7,7 @@ data = scrape()
 client = pymongo.MongoClient('mongodb://localhost:27017')
 db = client.mars_db
 collection = db.mars
-collection.drop()
-collection.insert_one(data)
-pprint(data)
+if data:
+    collection.drop()
+    collection.insert_one(data)
+    pprint(data)
